Frontend: Merge regex subroutine tests with regular ones
